requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.21.0
torch>=1.13.0
pyyaml>=6.0
tqdm
matplotlib
pandas